requirements.txt
asgiref==3.5.1
async-generator==1.10
attrs==21.4.0
autopep8==1.6.0
certifi==2022.9.24
colorama==0.4.5
coverage==6.5.0
Django==4.0.4
django-debug-toolbar==3.7.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
docopt==0.6.2
exceptiongroup==1.0.0rc9
Faker==13.12.0
flake8==4.0.1
h11==0.14.0
idna==3.4
iniconfig==1.1.1
mccabe==0.6.1
mypy==0.950
mypy-extensions==0.4.3
outcome==1.2.0
packaging==21.3
parameterized==0.8.1
Pillow==9.1.1
pluggy==1.0.0
py==1.11.0
pycodestyle==2.8.0
pyflakes==2.4.0
PyJWT==2.6.0
pyparsing==3.0.9
PySocks==1.7.1
pytest==7.1.2
pytest-django==4.5.2
pytest-watch==4.2.0
python-dateutil==2.8.2
python-dotenv==0.21.0
pytz==2022.5
selenium==4.5.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.2
toml==0.10.2
tomli==2.0.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.2.0
urllib3==1.26.12
watchdog==2.1.9
wsproto==1.2.0
